feat(groth16): required traits for `AffineCurve`
