+++ src/lib.rs
@@ -16,7 +16,7 @@ mod sign;
 
 use ark_ed_on_bls12_377::EdwardsProjective;
 pub use constants::ZETA;
-pub use element::Element;
+pub use element::{AffineElement, Element};
 pub use encoding::Encoding;
 pub use error::EncodingError;
 pub use field_ext::FieldExt;
