+++ src/ops/projective.rs
@@ -2,10 +2,6 @@ use core::ops::{Add, AddAssign, Mul, MulAssign, Neg, Sub, SubAssign};
 
 use crate::{Element, Fr};
 
-////////////////////////////////////////////////////////////////////////////////
-// Group operations: Projective
-////////////////////////////////////////////////////////////////////////////////
-
 impl<'a, 'b> Add<&'b Element> for &'a Element {
     type Output = Element;
